Fix browser-test
tape-run2 is a now outdated fork of tape-run intended to circumvent
a bug that doesn't exist anymore. Also phantomjs was not included as
a dev dependency, so npm run browser-test did not work out of the box.
Switching back to tape-run brings in electron as the default test
browser, so use that instead.
One of the tests in simple.js used fs methods which doesn't work in
the browser test.